Split out syntax module
This commit tries to do a "pure rename" part,
refactor will come next.
